Structure AlgoliaClient dependency
